requirements-dev.txt
pytest==8.3.5
pytest-django==4.11.1
pytest-cov==6.1.1
